% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{analyzeMassSpectralImages}
\alias{analyzeMassSpectralImages}
\title{Use image analysis to obtain csv of mass spectrum from an image.}
\usage{
analyzeMassSpectralImages(image_directory_in_path)
}
\arguments{
\item{image_directory_in_path}{The directory containing the images to be read.}
}
\description{
Read a png with left-most pixel low mass, right-most pixel high mass
Filename: reference~compoundName_lowmz-rightmz
}
\examples{
analyzeMassSpectralImages()
}
