% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{buildPolylist}
\alias{buildPolylist}
\title{Build a human-readable matrix of samples, analytes, and measurements}
\usage{
buildPolylist(
  samples_monolist_in_path,
  analytes_monolist_in_path,
  measurements_monolist_in_path = NULL,
  centroid = "~~~",
  polylist_out_path
)
}
\arguments{
\item{samples_monolist_in_path}{The monolist of samples to be used in building the polylist. Should contain a column "sample_unique_ID".}

\item{analytes_monolist_in_path}{The monolist of analytes to be used in building the polylist. Should contain a column "analyte_unique_ID".}

\item{measurements_monolist_in_path}{The monolist of measurements to be used in building the polylist. Should contain columns "sample_unique_ID", "analyte_unique_ID", and a column of values corresponding to measurements}

\item{centroid}{The characters to use in the centroid}

\item{polylist_out_path}{The path to which the polylist should be written}
}
\description{
Build a human-readable matrix of samples, analytes, and measurements
}
\keyword{lists}
