% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{buildTree}
\alias{buildTree}
\title{Construct various types of phylogenetic trees from alignments or other trees}
\usage{
buildTree(
  scaffold_type = c("amin_alignment", "nucl_alignment", "newick"),
  scaffold_in_path,
  members = NULL,
  gblocks = FALSE,
  gblocks_path = NULL,
  ml = FALSE,
  model_test = FALSE,
  bootstrap = FALSE,
  rois = FALSE,
  rois_data = NULL,
  ancestral_states = FALSE,
  root = NULL
)
}
\arguments{
\item{scaffold_type}{The type of information that should be used to build the tree. One of "amin_alignment", "nucl_alignment", or "newick"}

\item{scaffold_in_path}{The path to the information that should be used to build the tree.}

\item{members}{The tips of the tree that should be included. Default is to include everything.}

\item{gblocks}{TRUE/FALSE whether to use gblocks on the alignment}

\item{gblocks_path}{Path to the gblocks module, perhaps something like "/Users/lucasbusta/Documents/Science/_Lab_Notebook/Bioinformatics/programs/Gblocks_0.91b/Gblocks"}

\item{ml}{TRUE/FALSE whether to use maximum likelihood when constructing the tree.}

\item{model_test}{TRUE/FALSE whether to test various maximum likelihood models while constructing the tree}

\item{bootstrap}{TRUE/FALSE whether to calculate bootstrap values for tree nodes.}

\item{rois}{TRUE/FALSE whether to use only certain portions of an alignment for constructing the tree}

\item{rois_data}{The position in the alignment to use when constructing the tree. Default = NULL, i.e. all positions.}

\item{ancestral_states}{TRUE/FALSE whether to calculate ancestral states at nodes in the tree. Requires specifying a root via the 'root' parameter}

\item{root}{The tree tip to use as the root of the tree}
}
\description{
Construct various types of phylogenetic trees from alignments or other trees
}
