% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{convertCDFstoCSVs}
\alias{convertCDFstoCSVs}
\title{Convert mass spectral datafiles (CDF) into a csv file}
\usage{
convertCDFstoCSVs(
  paths_to_cdfs,
  min_mz = 50,
  max_mz = 800,
  min_rt = NULL,
  max_rt = NULL,
  force = FALSE
)
}
\arguments{
\item{paths_to_cdfs}{Paths to CDF files}

\item{min_mz}{Smallest m/z value to process (acts like a filter)}

\item{max_mz}{Largest m/z value to process (acts like a filter)}

\item{min_rt}{(optional) Smallest retention time to process (acts like a filter)}

\item{max_rt}{(optional) Largest retention time to process (acts like a filter)}

\item{force}{Convert all CDF files, even if they've already been converted previously}
}
\description{
Convert mass spectral datafiles (CDF) into a csv file
}
