% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{drawAlignment}
\alias{drawAlignment}
\title{Generate a multiple alignment graphic using ggplot}
\usage{
drawAlignment(infile, seqlist, alignment_labels, wrap = FALSE,
  wrap_length = NULL, roi = FALSE, roi_data = NULL,
  consensus = FALSE, consensus_height = 5, funct_assoc = FALSE,
  funct_assoc_data = NULL, funct_assoc_height = 5,
  highlights = FALSE, highlights_data = NULL, hlines = FALSE,
  hlines_data = NULL, tick_spacing = 5, ticks_text_size = 30,
  order = NULL, color_pal = NULL)
}
\arguments{
\item{infile}{The alignment to use (fasta file)}

\item{seqlist}{A dataframe with metadata for the alignment}

\item{alignment_labels}{The name of the column to label the entries in the alignment with}

\item{wrap}{TRUE/FALSE whether to use a column of plots to show the alignment}

\item{wrap_length}{Length at which to cut the wrap}

\item{roi}{TRUE/FALSE whether to only show regions of interest in the alignment}

\item{roi_data}{Dataframe defining the regions to show}

\item{consensus}{TRUE/FALSE whether to include a consensus line in the alignment}

\item{consensus_height}{Height of the consensus line plot}

\item{funct_assoc}{TRUE/FALSE whether to show functional associations in the alignment}

\item{funct_assoc_data}{Columns in the metadata dataframe to use in searching for functional association (e.g. "c(sterol~other, sterol~cyclo)")}

\item{funct_assoc_height}{Height of the functional association lines}

\item{highlights}{TRUE/FALSE whether to highlight certain sites in the alignment}

\item{highlights_data}{Dataframe specifying which site to highlight}

\item{hlines}{TRUE/FALSE whether to divide the alignment with horizontal lines}

\item{hlines_data}{Dataframe specifying where to divide alignment}

\item{tick_spacing}{Spacing between x-axis ticks}

\item{ticks_text_size}{Size of x-axis ticks text}

\item{order}{Order in which to display the members of the alignment}

\item{color_pal}{Color palette to use}
}
\description{
Allows the user to plot a multiple alignent using ggplot's grammar of graphics
}
\examples{
drawAlignment()
}
