% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{drawMolecules}
\alias{drawMolecules}
\title{Draw chemical structures from CSV}
\usage{
drawMolecules(csv_in_path)
}
\arguments{
\item{csv_in_path}{Path to the csv with the molecular coordinates.}
}
\description{
Draw chemical structures from CSV
}
\examples{
drawMolecules()
}
