% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{extractORFs}
\alias{extractORFs}
\title{Extract open reading frames from multifasta files}
\usage{
extractORFs(file, write_out_ORFs = FALSE, overwrite = FALSE)
}
\arguments{
\item{file}{The multifasta file to analyze}

\item{write_out_ORFs}{TRUE/FALSE whether to write out a new fasta that contains the ORFs}

\item{overwrite}{TRUE/FALSE Optionally overwrite the input file with the ORF file}
}
\description{
Extract open reading frames from multifasta files
}
