% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{generateTicks}
\alias{generateTicks}
\title{Create major and minor axes ticks and labels}
\usage{
generateTicks(major_ticks, minor_freq = 4, major_tick_size,
  minor_tick_size)
}
\arguments{
\item{major_ticks}{Values at which major ticks should be generated}

\item{minor_freq}{Number of minor ticks between each major tick}
}
\description{
Create major and minor axes ticks and labels
}
