% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{integrationAppLite}
\alias{integrationAppLite}
\title{A Shiny app to integrate GC-FID and GC-MS data}
\usage{
integrationAppLite(
  CDF_directory_path,
  create_new_samples_monolist = FALSE,
  create_new_peak_monolist = FALSE,
  zoom_and_scroll_rate = 100,
  baseline_window = 400
)
}
\arguments{
\item{zoom_and_scroll_rate}{Defines intervals of zooming and scrolling movement while running the app}

\item{chromatograms}{A data frame containing columns: "rt", "tic", and "path_to_cdf_csv", which contain retention time, total ion chromatogram intensities, and paths to CDF.csv files generated by the convertCDFstoCSVs function.}

\item{x_axis_start}{A numeric value for the lower x-axis bounds on the plot generated by the app. Defaults to full length.}

\item{x_axis_end}{A numeric value for the upper x-axis bounds on the plot generated by the app. Defaults to full length.}

\item{samples_monolist_path}{A path to a .csv file containing metadata for the samples you wish to analyze. Requied columns are: "rt_offset", "baseline_window", and "path_to_cdf_csv", which are for aligning chromatograms, adjusting baseline determination, and defining the path to the CDF.csv files for each sample, respectively.}

\item{samples_monolist_subset}{Optional, a numeric vector (for example, "c(1:10)"), defining a subset of samples to be loaded.}

\item{peaks_monolist_path}{A path to a .csv file containing metadata for all peaks in the sample set. Required columns are: peak_start", "peak_end", "path_to_cdf_csv", "area", "peak_number_within_sample", "rt_offset", "peak_start_rt_offset", "peak_end_rt_offset". This file is automatically generated by the app.}
}
\description{
A Shiny app to integrate GC-FID and GC-MS data
}
