% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{mergePeakLists}
\alias{mergePeakLists}
\title{Merge multiple peaklists}
\usage{
mergePeakLists(analysis_directory_path)
}
\arguments{
\item{analysis_directory_path}{Path to the directory above all the peak lists}
}
\description{
Merge multiple peaklists
}
