% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{OsDirectoryPathCorrect}
\alias{OsDirectoryPathCorrect}
\title{OS-aware correction of directory paths}
\usage{
OsDirectoryPathCorrect(directory_path)
}
\arguments{
\item{directory_in_path}{The path to correct}
}
\description{
OS-aware correction of directory paths
}
