% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{pGroups}
\alias{pGroups}
\title{Determine which samples are identical based on p values}
\usage{
pGroups(data)
}
\arguments{
\item{data}{}
}
\description{
Determine which samples are identical based on p values
}
