% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{phylogeneticSignal}
\alias{phylogeneticSignal}
\title{Convert mass spectral datafiles (CDF) into a csv file}
\usage{
phylogeneticSignal(trait, tree, replicates = 999, cost = NULL)
}
\arguments{
\item{trait}{A vector of the variable, named according to which tree tip it comes from}

\item{tree}{A phylogenetic tree with tips that exactly match the names of trait}

\item{replicates}{Number of random replications to run}

\item{cost}{Optional. A specialized transition matrix}
}
\description{
Convert mass spectral datafiles (CDF) into a csv file
}
