% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{pruneTree}
\alias{pruneTree}
\title{Prune a tree so it only contains user-specified tips}
\usage{
pruneTree(tree_in_path, tips_to_keep)
}
\arguments{
\item{tips_to_keep}{The tips to keep}

\item{tree}{The tree to manipulate}
}
\description{
Prune a tree so it only contains user-specified tips
}
