% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readChromatograms}
\alias{readChromatograms}
\title{Import chromatograms stored as ChemStation exported .csv files}
\usage{
readChromatograms(dir, normalize_level = c("none", "var1", "var2"),
  normalize_range = NULL)
}
\arguments{
\item{dir}{A directory containing (ONLY) the chromatogram(s) that are to be plotted}

\item{normalize_level}{Level at which to normalize chromatogram}

\item{normalize_range}{y range to consider during normalization. Useful for normalizing but excluding solvent peak(s), for example}
}
\description{
Allows the user to import chromatograms stored as .csv files. Files should have originated from an Agilent GC system running ChemStation
Files should be in the format of character~var1-var2.csv
"character" is the name of the chromatogram, var1 and var2 are variables you want associated with that chromatogram.
}
\examples{
readChromatograms()
}
