% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readCSV}
\alias{readCSV}
\title{Interactive selection of a CSV file to read}
\usage{
readCSV(sep = c(",", " "))
}
\arguments{
\item{sep}{The delimiter to use when writing the file out. Default is a comma, i.e. CSV.}
}
\description{
Interactive selection of a CSV file to read
}
