% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readFasta}
\alias{readFasta}
\title{Returns contents of one fasta file as a StringSet object.}
\usage{
readFasta(fasta_in_path, fasta_type = "nonspecific")
}
\arguments{
\item{fasta_in_path}{The path to the fasta to read}

\item{fasta_type}{The type of sequence data contained in the fasta file. Options: "nonspecific", "DNA", "RNA", or "AA"}
}
\description{
Returns contents of one fasta file as a StringSet object.
}
