% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readGFFs}
\alias{readGFFs}
\title{readGFFs}
\usage{
readGFFs(input_frame, subset_mode = c("none", "goi_only", "foi_only",
  "goi_region", "foi_region", "name_check"), goi = NULL, foi = NULL,
  region_reach = 2e+05, concatenate_by_species = TRUE,
  concatenation_spacing = 10000, omit = NULL)
}
\arguments{
\item{input_frame}{A dataframe with columns: Genus_species, GFF_in_path, region_reach, concatenation_spacing}

\item{query}{A .fa file containing the query}

\item{phylochem}{The phylochem object to which the data should be bound}
}
\description{
Read multiple genome feature files into a tidy dataframe
}
