% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readManyFasta}
\alias{readManyFasta}
\title{Returns the contents of many fasta files as a StringSet object.}
\usage{
readManyFasta(fasta_in_paths, fasta_type = "nonspecific")
}
\arguments{
\item{fasta_in_paths}{The paths to the fastas to read}

\item{fasta_type}{The type of sequence data contained in the fasta file. Options: "nonspecific", "DNA", "RNA", or "AA"}
}
\description{
Returns the contents of many fasta files as a StringSet object.
}
