% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readMonolist}
\alias{readMonolist}
\title{Reads a monolist}
\usage{
readMonolist(monolist_in_path)
}
\arguments{
\item{monolist_in_path}{The path to the monolist (in .csv format) to be read}
}
\description{
Reads a monolist
}
