% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readPolylist}
\alias{readPolylist}
\title{Reads a polylist}
\usage{
readPolylist(
  polylist_in_path,
  centroid = "~~~",
  table_value_unit = "abundance"
)
}
\arguments{
\item{polylist_in_path}{The path to the polylist (in .csv format) to be read}

\item{centroid}{The characters in the cell that defines the boundaries of the multiple row and column headers}

\item{table_value_unit}{The name to be given to the column that will contain the values in the polylist table}
}
\description{
Reads a spreadsheet in polylist format (wide format, multiple column and row headers), and converts it into tidy format
}
