% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readSpectra}
\alias{readSpectra}
\title{Import mass spectral data}
\usage{
readSpectra(dir)
}
\arguments{
\item{dir}{The directory containing the mass spectral (.csv) files of interest}
}
\description{
Used to import one or more mass spectra and collect the data in a dataframe.
The names of the files containing the mass spectra must have the following format: "var2~var3-var4.csv", with var1 being the file's position in the directory
}
\examples{
readSpectra()
}
