% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{readTree}
\alias{readTree}
\title{Reads a phylogenetic tree}
\usage{
readTree(tree_in_path)
}
\arguments{
\item{tree_in_path}{The path to the phylogenetic tree}
}
\description{
Reads a phylogenetic tree
}
