% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{runMatrixAnalysis}
\alias{runMatrixAnalysis}
\title{Runs a matrix analysis (clustering, kmeans, pca).}
\usage{
runMatrixAnalysis(
  data,
  analysis = c("hclust", "pca", "pca-ord", "pca-dim"),
  column_w_names_of_multiple_analytes = NULL,
  column_w_values_for_multiple_analytes = NULL,
  columns_w_values_for_single_analyte = NULL,
  columns_w_additional_analyte_info = NULL,
  columns_w_sample_ID_info = NULL,
  transpose = FALSE,
  unknown_sample_ID_info = NULL,
  scale_variance = TRUE,
  kmeans = c("none", "auto", "elbow", "1", "2", "3", "etc."),
  na_replacement = c("none", "mean", "zero", "drop"),
  output_format = c("wide", "long"),
  ...
)
}
\arguments{
\item{data}{The data frame or tibble to analyze}

\item{analysis}{The type of analysis to run. Can be one of: "hclust" (heirarchical clustering), "pca" (principal components analysis), "pca-ord" (principal components analysis ordination plot), or "pca-dim" (principal components scree plot)}

\item{column_w_names_of_multiple_analytes}{}
}
\description{
Runs a matrix analysis (clustering, kmeans, pca).
}
