% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writeCSV}
\alias{writeCSV}
\title{Interactive selection of a CSV file to write}
\usage{
writeCSV(monolist)
}
\arguments{
\item{monolist}{The monolist to write out}
}
\description{
Interactive selection of a CSV file to write
}
