% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writeFasta}
\alias{writeFasta}
\title{Writes one or more sequences as a single fasta file}
\usage{
writeFasta(XStringSet, fasta_out_path)
}
\arguments{
\item{XStringSet}{The sequences to write out}

\item{fasta_out_path}{The path to where the file should be written}
}
\description{
Writes one or more sequences as a single fasta file
}
