% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writeMonolist}
\alias{writeMonolist}
\title{Writes a monolist}
\usage{
writeMonolist(monolist, monolist_out_path)
}
\arguments{
\item{monolist_out_path}{The path to where the monolist should be written}
}
\description{
Writes a monolist
}
