% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writePolylist}
\alias{writePolylist}
\title{Writes a polylist}
\usage{
writePolylist(polylist, polylist_out_path)
}
\arguments{
\item{polylist_out_path}{The path to where the polylist should be written}
}
\description{
Writes a polylist
}
