% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writeSingleFastas}
\alias{writeSingleFastas}
\title{Writes one or more sequences as individual fasta file(s), each with one sequence}
\usage{
writeSingleFastas(XStringSet, fasta_out_directory_path)
}
\arguments{
\item{XStringSet}{The sequence(s) to write out}

\item{fasta_out_directory_path}{The path to the directory where the file(s) should be written}
}
\description{
Writes one or more sequences as individual fasta file(s), each with one sequence
}
