% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writeSupplementalTable}
\alias{writeSupplementalTable}
\title{Writes a human-readable supplemental table of quantitative data}
\usage{
writeSupplementalTable(supplementalTable, round_level, file_out_path,
  replicates)
}
\arguments{
\item{supplementalTable}{The data to write out}

\item{round_level}{The number of decimal places summary statistics should have}

\item{file_out_path}{The path to the location where the file should be written}
}
\description{
Writes a human-readable supplemental table of quantitative data
}
