% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{writeTree}
\alias{writeTree}
\title{Writes a phylogenetic tree}
\usage{
writeTree(tree, tree_out_path)
}
\arguments{
\item{tree_out_path}{The path to where the tree should be written}
}
\description{
Writes a phylogenetic tree
}
